@@ -4,7 +4,7 @@
     "description": "A starter project for Laravel apps using Inertia and Vue.",
     "license": "MIT",
     "require": {
-        "php": "^8.2",
+        "php": "^8.4",
         "filament/filament": "^4.0",
         "inertiajs/inertia-laravel": "^2.0",
         "laravel/framework": "^12.0",
@@ -20,6 +20,7 @@
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.13",
+        "driftingly/rector-laravel": "^2.1",
         "fakerphp/faker": "^1.23",
         "laravel/boost": "^1.8",
         "laravel/pint": "^1.17",
@@ -28,6 +29,7 @@
         "pestphp/pest": "^3.0",
         "pestphp/pest-plugin-faker": "^3.0",
         "pestphp/pest-plugin-laravel": "^3.0",
+        "rector/rector": "^2.2",
         "spatie/laravel-ray": "^1.37"
     },
     "autoload": {
@@ -57,6 +59,9 @@
         ],
         "build:test": [
             "composer install --no-interaction"
+        ],
+        "rector": [
+            "./vendor/bin/rector"
         ],
         "lint": [
             "./vendor/bin/pint --test"
