@@ -5,9 +5,7 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('health', Spatie\Health\Http\Controllers\HealthCheckResultsController::class)->middleware(['auth', 'role:'.Role::SUPER_ADMIN->value]);
 
-Route::get('elements', function () {
-    return inertia('Elements');
-})->middleware(['auth', 'role:'.Role::SUPER_ADMIN->value])->name('elements');
+Route::get('elements', fn () => inertia('Elements'))->middleware(['auth', 'role:'.Role::SUPER_ADMIN->value])->name('elements');
 
 Route::controller(App\Http\Controllers\RegisterController::class)
     ->middleware(['guest'])
