@@ -34,6 +34,6 @@ class RegisterController extends Controller
 
         event(new Registered($user));
 
-        return redirect()->route('home');
+        return to_route('home');
     }
 }
